requirements.txt
fasteners==0.15
google-apitools==0.5.31
httplib2==0.18.1
monotonic==1.5
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
rsa==4.1
six==1.15.0
google-businesscommunications==1.0.0